+++ packages/firebase_dart/lib/src/database/impl/treestructureddata/name.dart
@@ -124,3 +124,7 @@ class _NameImpl extends Name {
     return null;
   }
 }
+
+extension NamePathX on Path<Name> {
+  String asString() => join('/');
+}